+++ myTrip/static/src/profile/profile.less
@@ -31,3 +31,7 @@
 	font-size: 2.8em;
 
 }
+
+.floatingLabelText {
+	color: blue;
+}
